% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{adjust_ts_data}
\alias{adjust_ts_data}
\title{Adjust \code{ts_data}}
\usage{
adjust_ts_data(data)
}
\arguments{
\item{data}{Matrix or data.frame to adjust.}
}
\value{
An adjusted \code{ts_data}.
}
\description{
Convert a compatible dataset to a \code{ts_data} object by setting
column names, class, and the \code{sw} attribute consistently.
}
