% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal_oversampling.R
\name{bal_oversampling}
\alias{bal_oversampling}
\title{Class Balance Oversampling}
\usage{
bal_oversampling(attribute)
}
\arguments{
\item{attribute}{The attribute to be balanced through oversampling.}
}
\value{
An instance of the BalanceOversampling class.
}
\description{
The R class BalanceOversampling represents a method to balance the class distribution of a dataset by oversampling the minority class. It can be used to increase the representation of the minority class in the data used for modeling or analysis.
}
\examples{
data(iris)
mod_iris <- iris[c(1:50,51:71,101:111),]

bal <- bal_oversampling('Species')
bal <- daltoolbox::fit(bal, mod_iris)
adjust_iris <- daltoolbox::transform(bal, mod_iris)
table(adjust_iris$Species)
}
