% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal_subsampling.R
\name{bal_subsampling}
\alias{bal_subsampling}
\title{Class Balance Subsampling}
\usage{
bal_subsampling(attribute)
}
\arguments{
\item{attribute}{The attribute to be balanced through subsampling.}
}
\value{
An instance of the BalanceSubsampling class.
}
\description{
The R class BalanceSubsampling represents a method to balance the class distribution of a dataset by subsampling the majority class. It can be used to reduce the representation of the majority class in the data used for modeling or analysis.
}
\examples{
data(iris)
mod_iris <- iris[c(1:50,51:71,101:111),]

bal <- bal_subsampling('Species')
bal <- daltoolbox::fit(bal, mod_iris)
adjust_iris <- daltoolbox::transform(bal, mod_iris)
table(adjust_iris$Species)
}
