% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_balance.R
\name{balance_oversampling}
\alias{balance_oversampling}
\title{Class Balance Oversampling}
\usage{
balance_oversampling(attribute)
}
\arguments{
\item{attribute}{The attribute to be balanced through oversampling.}
}
\value{
An instance of the BalanceOversampling class.
}
\description{
The R class BalanceOversampling represents a method to balance the class distribution of a dataset by oversampling the minority class. It can be used to increase the representation of the minority class in the data used for modeling or analysis.
}
\details{
The BalanceOversampling class has the following properties:
data: the oversampled data frame;
seed: the random seed used for reproducibility.
The BalanceOversampling class has the following methods:
summary(): provides a summary of the oversampled dataset, including the class distribution before and after oversampling;
plot(): produces a plot of the class distribution before and after oversampling.
}
