% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs.R
\name{cla_fs}
\alias{cla_fs}
\title{Feature Selection}
\usage{
cla_fs(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
An instance of the FeatureSelection class.
}
\description{
Feature selection is a process of selecting a subset of relevant features from a larger set of features in a dataset for use in model training. The FeatureSelection class in R provides a framework for performing feature selection.
}
\examples{
#See ?cla_fs_fss for an example of feature selection
}
