% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs_fss.R
\name{cla_fs_fss}
\alias{cla_fs_fss}
\title{Forward Stepwise Selection}
\usage{
cla_fs_fss(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
An instance of the ForwardStepwiseSelection class.
}
\description{
Forward stepwise selection is a technique for feature selection in which features are added to a model one at a time, based on their ability to improve the performance of the model. The ForwardStepwiseSelection class in R provides a framework for performing forward stepwise selection.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(cla_fs_fss("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
