% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs_ig.R
\name{cla_fs_ig}
\alias{cla_fs_ig}
\title{Information Gain}
\usage{
cla_fs_ig(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
A instance of \code{cla_fs_ig}
}
\description{
Information Gain is a feature selection technique used in machine learning to determine the relevance of a feature to the target variable. It measures the amount of information obtained for the target variable by knowing the presence or absence of a feature.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(cla_fs_ig("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
