% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs_lasso.R
\name{cla_fs_lasso}
\alias{cla_fs_lasso}
\title{Feature Selection using Lasso regression}
\usage{
cla_fs_lasso(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
An instance of the FeatureSelectionLasso class.
}
\description{
Feature selection using Lasso regression is a technique for selecting a subset of relevant features from a larger set of features in a dataset for use in model training. The FeatureSelectionLasso class in R provides a framework for performing feature selection using Lasso regression.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(cla_fs_lasso("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
