% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs_relief.R
\name{cla_fs_relief}
\alias{cla_fs_relief}
\title{Relief}
\usage{
cla_fs_relief(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\description{
The Relief algorithm is a feature selection technique used in machine learning to determine the relevance of a feature to the target variable. It calculates the relevance of a feature by considering the difference in feature values between nearest neighbors of the same and different classes.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(cla_fs_relief("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
