% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{do_fit}
\alias{do_fit}
\title{Fit Time Series Model}
\usage{
do_fit(obj, x, y = NULL)
}
\arguments{
\item{obj}{Model object to be fitted.}

\item{x}{Matrix or data.frame with input features.}

\item{y}{Vector or matrix with target values.}
}
\value{
A fitted object (same class as \code{obj}).
}
\description{
Generic for fitting a time series model.
Descendants should implement \verb{do_fit.<class>}.
}
