% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{do_predict}
\alias{do_predict}
\title{Predict Time Series Model}
\usage{
do_predict(obj, x)
}
\arguments{
\item{obj}{Fitted model object.}

\item{x}{Matrix or data.frame with input features to predict.}
}
\value{
Numeric vector with predicted values.
}
\description{
Generic for predicting with a fitted time series model.
Descendants should implement \verb{do_predict.<class>}.
}
