% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizers}
\alias{fertilizers}
\title{Fertilizers (Regression)}
\format{
list of fertilizers' time series.
}
\source{
This dataset was obtained from the MASS library.
}
\usage{
data(fertilizers)
}
\description{
List of Brazilian fertilizers consumption of N, P2O5, K2O.
\itemize{
\item brazil_n: nitrogen consumption from 1961 to 2020.
\item brazil_p2o5: phosphate consumption from 1961 to 2020.
\item brazil_k2o: potash consumption from 1961 to 2020.
}
}
\examples{
data(fertilizers)
head(fertilizers$brazil_n)
}
\references{
International Fertilizer Association (IFA): http://www.fertilizer.org.
}
\keyword{datasets}
