% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspredbench_loadfulldata.R
\name{loadfulldata}
\alias{loadfulldata}
\title{Load Full Dataset From Mini Data Object}
\usage{
loadfulldata(x)
}
\arguments{
\item{x}{A mini dataset object that contains \code{attr(x, "url")}.}
}
\value{
The full dataset object loaded from the remote \code{.RData} file.
}
\description{
Downloads and loads the full \code{.RData} object referenced by \code{attr(x, "url")}
from a mini dataset object loaded from \verb{data/}.
}
