% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_aug_none.R
\name{ts_aug_none}
\alias{ts_aug_none}
\title{no augmentation}
\usage{
ts_aug_none()
}
\value{
a \code{ts_aug_none} object.
}
\description{
Does not make data augmentation.
}
\examples{
library(daltoolbox)
data(sin_data)

#convert to sliding windows
xw <- ts_data(sin_data$y, 10)

#no data augmentation
augment <- ts_aug_none()
augment <- fit(augment, xw)
xa <- transform(augment, xw)
ts_head(xa)
}
