% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_smooth.R
\name{ts_fil_smooth}
\alias{ts_fil_smooth}
\title{Time Series Smooth}
\usage{
ts_fil_smooth()
}
\value{
A \code{ts_fil_smooth} object.
}
\description{
Remove or reduce randomness (noise) using a robust smoothing
strategy that first mitigates outliers and then smooths residual variation.
}
\examples{
# time series with noise
library(daltoolbox)
data(tsd)
tsd$y[9] <- 2*tsd$y[9]

# filter
filter <- ts_fil_smooth()
filter <- fit(filter, tsd$y)
y <- transform(filter, tsd$y)

# plot
plot_ts_pred(y=tsd$y, yadj=y)
}
