% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_head}
\alias{ts_head}
\title{Extract the First Observations from a \code{ts_data} Object}
\usage{
ts_head(x, n = 6L, ...)
}
\arguments{
\item{x}{\code{ts_data} object}

\item{n}{number of rows to return}

\item{...}{optional arguments}
}
\value{
The first n observations of a \code{ts_data} (as a matrix/data.frame).
}
\description{
Return the first n observations from a \code{ts_data}.
}
\examples{
data(tsd)
data10 <- ts_data(tsd$y, 10)
ts_head(data10)
}
