% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_deterministic.R
\name{ts_persist}
\alias{ts_persist}
\title{Persistence Deterministic Predictor}
\usage{
ts_persist()
}
\value{
A \code{ts_persist} object, inheriting from \code{ts_deterministic}.
}
\description{
Forecast a univariate series by repeating its most recent
observed value.
}
\examples{
series <- c(10, 11, 11, 11)
model <- ts_persist()
model <- daltoolbox::fit(model, x = series)
predict(model, steps_ahead = 3)
}
