% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{ts_reg}
\alias{ts_reg}
\title{TSReg}
\usage{
ts_reg()
}
\value{
A \code{ts_reg} object (S3) to be extended by concrete models.
}
\description{
Base class for time series regression models that operate
directly on time series (non-sliding-window specialization).
}
\details{
This class is intended to be subclassed by modeling backends that
do not require the sliding-window interface. Methods such as \code{fit()},
\code{predict()}, and \code{evaluate()} dispatch on this class.
}
\examples{
# This is an abstract base class; use concrete subclasses.
}
