% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs_lasso.R
\name{cla_fs_lasso}
\alias{cla_fs_lasso}
\title{Feature Selection using Lasso}
\usage{
cla_fs_lasso(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
A \code{cla_fs_lasso} object.
}
\description{
Feature selection using Lasso regression is a technique for selecting a subset of relevant features.
It wraps the glmnet library.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(cla_fs_lasso("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
