% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla_fs_relief.R
\name{cla_fs_relief}
\alias{cla_fs_relief}
\title{Relief}
\usage{
cla_fs_relief(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\description{
The Relief algorithm is a feature selection technique used in machine learning to determine the relevance of a feature to the target variable. It calculates the relevance of a feature by considering the difference in feature values between nearest neighbors of the same and different classes.
}
\details{
The relief function has the following parameters: data: the data frame containing the features and target variable; target: the name of the target variable in the data frame; nn: the number of nearest neighbors to consider (optional); sample.size: the number of samples to use in the estimation of the feature distribution (optional). The relief function returns a named numeric vector of Relief scores for each feature in the data frame.
}
