% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reg.R
\name{R2.ts}
\alias{R2.ts}
\title{R2}
\usage{
R2.ts(actual, prediction)
}
\arguments{
\item{actual}{Numeric vector of observed values.}

\item{prediction}{Numeric vector of predicted values.}
}
\value{
Numeric scalar with R-squared.
}
\description{
Compute coefficient of determination (R-squared).
}
