% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_fft.R
\name{ts_fil_fft}
\alias{ts_fil_fft}
\title{FFT Filter}
\usage{
ts_fil_fft()
}
\value{
a \code{ts_fil_fft} object.
}
\description{
FFT Filter
}
\examples{
# time series with noise
library(daltoolbox)
data(tsd)
tsd$y[9] <- 2*tsd$y[9]

# filter
filter <- ts_fil_fft()
filter <- fit(filter, tsd$y)
y <- transform(filter, tsd$y)

# plot
plot_ts_pred(y=tsd$y, yadj=y)
}
