% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_none.R
\name{ts_fil_none}
\alias{ts_fil_none}
\title{no filter}
\usage{
ts_fil_none()
}
\value{
a \code{ts_fil_none} object.
}
\description{
Does not make data filter
}
\examples{
# time series with noise
library(daltoolbox)
data(tsd)
tsd$y[9] <- 2*tsd$y[9]

# filter
filter <- ts_fil_none()
filter <- fit(filter, tsd$y)
y <- transform(filter, tsd$y)

# plot
plot_ts_pred(y=tsd$y, yadj=y)
}
