% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_none.R
\name{ts_fil_none}
\alias{ts_fil_none}
\title{no filter}
\usage{
ts_fil_none()
}
\value{
a \code{ts_fil_none} object.
}
\description{
Does not make data filter
}
\examples{
library(daltoolbox)
data(sin_data)

#convert to sliding windows
xw <- ts_data(sin_data$y, 10)

#no data filter
filtering <- ts_fil_none()
filtering <- fit(filtering, xw)
xa <- transform(filtering, xw)
ts_head(xa)
}
