% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fil_smooth.R
\name{ts_fil_smooth}
\alias{ts_fil_smooth}
\title{Time Series Smooth}
\usage{
ts_fil_smooth()
}
\value{
a \code{ts_fil_smooth} object.
}
\description{
Used to remove or reduce randomness (noise).
}
\examples{
# time series with noise
library(daltoolbox)
data(sin_data)
sin_data$y[9] <- 2*sin_data$y[9]
# convert to sliding windows
ts <- ts_data(sin_data$y, 10)
ts_head(ts, 3)
summary(ts[,10])

# filter
filter <- ts_fil_smooth()
filter <- fit(filter, sin_data$y)
y <- transform(filter, sin_data$y)

# plot
plot_ts_pred(y=sin_data$y, yadj=y)
}
