% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsd}
\alias{tsd}
\title{Time series example dataset}
\format{
\code{data.frame}.
}
\source{
This dataset was generated for examples.
}
\usage{
data(tsd)
}
\description{
Synthetic dataset based on a sine function.
\itemize{
\item x: correspond time from 0 to 10.
\item y: dependent variable for time series modeling.
}
}
\examples{
data(tsd)
head(tsd)
}
\keyword{datasets}
